Initial implementation of the GSI provider
This provider allows Terraform AWS user to manage GSI on dynamo table as
a separate resource as a workaround for
https://github.com/hashicorp/terraform-provider-aws/issues/671.

Using this new resource requires ignoring `global_secondary_index` and
`attribute` and the aws_dynamodb_table resource as described in the
example terraform file.

This provder also provides an `auto_import` mode which should make the
process of migration more seamless.